+++ project/geosource/tests/test_tasks.py
@@ -10,6 +10,8 @@ from project.geosource.tasks import run_model_object_method
 from project.geosource.tests.helpers import get_file
 
 
+@mock.patch("elasticsearch.client.IndicesClient.create")
+@mock.patch("elasticsearch.client.IndicesClient.delete")
 class TaskTestCase(TestCase):
     @classmethod
     def setUpTestData(cls):
@@ -21,7 +23,7 @@ class TaskTestCase(TestCase):
         )
         cls.element.groups.add(cls.group)
 
-    def test_task_refresh_data_method(self):
+    def test_task_refresh_data_method(self, mocked_es_delete, mocked_es_create):
         run_model_object_method.apply(
             (
                 self.element._meta.app_label,
@@ -35,7 +37,9 @@ class TaskTestCase(TestCase):
         self.assertEqual(Feature.objects.first().properties, {"id": 1, "test": 5})
         self.assertEqual(Layer.objects.first().authorized_groups.first().name, "Group")
 
-    def test_task_refresh_data_method_wrong_pk(self):
+    def test_task_refresh_data_method_wrong_pk(
+        self, mocked_es_delete, mocked_es_create
+    ):
         logging.disable(logging.WARNING)
         run_model_object_method.apply(
             (
@@ -47,7 +51,7 @@ class TaskTestCase(TestCase):
         )
         self.assertEqual(Layer.objects.count(), 0)
 
-    def test_task_wrong_method(self):
+    def test_task_wrong_method(self, mocked_es_delete, mocked_es_create):
         logging.disable(logging.ERROR)
         run_model_object_method.apply(
             (
@@ -60,7 +64,9 @@ class TaskTestCase(TestCase):
         self.assertEqual(Layer.objects.count(), 0)
 
     @mock.patch("project.geosource.models.Source.objects")
-    def test_task_good_method_error(self, mock_source):
+    def test_task_good_method_error(
+        self, mocked_es_delete, mocked_es_create, mock_source
+    ):
         mock_source.get.side_effect = ValueError
         logging.disable(logging.ERROR)
         run_model_object_method.apply(