+++ docs/source/others/changelog.rst
@@ -103,6 +103,14 @@ Changelog
 - Add more options to info content editor in config dashboard
 
 
+XXXX.X.X       (XXXX-XX-XX)
+---------------------------
+
+**Performance**
+
+- Optimize source refresh and ES indexation process -p
+
+
 2023.6.9       (2023-06-21)
 ---------------------------